content/base/test/test_bug562169-2.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=562169
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 562169</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=562169">Mozilla Bug 562169</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 562169 **/
/** Test that a newly created element has ltr directionality **/

ok(document.createElement("div").mozMatchesSelector(":-moz-dir(ltr)"),
   "Element should be ltr on creation");

</script>
</pre>
</body>
</html>
